+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.1.0
+# 0.2.0
 
-* Adds Android support
+* Adds iOS support for the instance manager.
+
+# 0.1.0
+
+* Adds Android support for the instance manager.